energyPoint/checkin

+++ oas-server/oas/src/main/java/com/cascv/oas/server/energy/model/EnergySourcePower.java
@@ -11,10 +11,10 @@ public class EnergySourcePower {
     @Getter @Setter Integer sourceCode;
     @Getter @Setter String sourceName;
     @Getter @Setter String type;
-    @Getter @Setter Integer powerSingle;
-    @Getter @Setter Integer powerTotal;
+    @Getter @Setter BigDecimal powerSingle;
+    @Getter @Setter BigDecimal powerTotal;
     @Getter @Setter BigDecimal powerIncreaseSpeed;
     @Getter @Setter String powerIncreaseSpeedUnitName;
-    @Getter @Setter Integer powerIncreaseSpeedUnitValue;
-    @Getter @Setter Integer powerMaxEachBall;
+    @Getter @Setter BigDecimal powerIncreaseSpeedUnitValue;
+    @Getter @Setter BigDecimal powerMaxEachBall;
 }
